% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ancestors.R
\name{ancestryR}
\alias{ancestryR}
\title{Helper function with additional recursive parameters}
\usage{
ancestryR(id, steps, siblings = FALSE, from = NULL, verbose = FALSE)
}
\arguments{
\item{id}{identifier given to a mathematician in the Mathematics Genealogy Project.}

\item{steps}{integer number of number of steps to follow from the root}

\item{siblings}{Boolean specifying the extent of the scrape. Are academic siblings to be reported as well?}

\item{from}{Math Gen id of the mathematician to follow upwards}

\item{verbose}{Boolean - should results be reported during the scraping? defaults to FALSE.}
}
\description{
Helper function with additional recursive parameters
}

