% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.R
\name{descendants}
\alias{descendants}
\title{Descendants of a Mathematician}
\usage{
descendants(id, steps = 3, includeself = TRUE)
}
\arguments{
\item{id}{identifier given to a mathematician in the Mathematics Genealogy Project.}

\item{steps}{integer if specified, the number of steps to follow from the root. If NULL, all descendants will be reported.}

\item{includeself}{Boolean, should the person him/herself be included in the results?}
}
\value{
data frame
}
\description{
Descendants of a Mathematician
}
\examples{
\dontrun{
au <- descendants(45024) # Antony Unwin
}
}

