% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ancestors.R
\name{ancestry}
\alias{ancestry}
\title{Get a set number of ancestors based on name and Math Gen ID}
\usage{
ancestry(id, steps, verbose = FALSE, includeself = TRUE)
}
\arguments{
\item{id}{identifier given to a mathematician in the Mathematics Genealogy Project.}

\item{steps}{integer number of number of steps to follow from the root}

\item{verbose}{Boolean - should results be reported during the scraping? defaults to FALSE.}
}
\value{
data frame with the following variables
\itemize{
  \item Name name of the mathematician
  \item mgID Math Gen ID
  \item School university granting the degree
  \item Thesistitle
  \item Degree
  \item Year
  \item advisorName
  \item advisorMGID Math Gen ID of the advisor(s)
}
}
\description{
Starting from the given node, doctoral advisors are followed recursively for the specified number of steps. In case no advisor is known, the algorithm stops even if the number of steps is not yet reached.
Loops in the academic tree (thicket??) are problematic - e.g. if a person was advised by two advisors, and one of them was an advisor to the other (advised by both academic brother and father). This results in duplicate entries.
Each call to ancestry results in multiple requests to the Mathematics Genealogy Project, so use responsibly.
}
\examples{
\dontrun{
hw <- ancestry(id = 145799, steps = 2) # Hadley Wickham
ancestry(id = 145799,  steps = 2, siblings = TRUE) # Hadley Wickham
dh <- ancestry(id=7298,  steps = 5) # David Hilbert

library(ggplot2)
library(geomnet)

ggplot() +
  geom_net(aes(from_id=factor(advisorMGID), to_id=factor(mgID),
               label=advisorName), directed=TRUE, data=dh) +
  theme_net() + xlim(c(-0.1, 1.1))
qplot(y = rep(1:10, length=nrow(dh)), x=as.numeric(as.character(Year)), label = Name,
      data=dh, geom="label", alpha = I(0.5))
}
}

