% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ancestors.R
\name{getNode}
\alias{getNode}
\title{Scrape data for an academic record}
\usage{
getNode(id, name)
}
\description{
Scrape data for an academic record
}
\examples{
mathGen:::getNode(id = 15860, "John Wilder Tukey")
}

